backend/requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-community
langchain-openai
langchain-google-genai

chromadb
faiss-cpu
pypdf
docx2txt
